Fix omnifunc interactions
Some hack-ish things, but seems to work. The findstart
behavior for issues/user completion is cleaned up, at
least, so that is more reliable.

Example situation that would break it previously:

Given:
  Foo
Do:
  I#

It should give you ALL issues, but it would previously
filter on that `F`. That should now be fixed